# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/user/users": {
      "get": {
        "tags": ["users"],
        "summary": "Obtenir tous les utilisateurs",
        "description": "Permet d'obtenir les informations de tous les utilisateurs de la base",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/user/create": {
      "post": {
        "tags": ["users"],
        "summary": "Créer un utilisateur",
        "description": "Permet d'ajouter un nouvelle utilisateur",
        "parameters": [
          {
            "name": "firstname",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "Jean"
          },
          {
            "name": "lastname",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "BON"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "required": true,
            "uniqueItems": true,
            "default": "test@test.com"

          },
          {
            "name": "password",
            "in": "formData",
            "type": "string",
            "required": true,
            "schema": {
              "pattern": "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[a-zA-Z]).{8,}$"
            },
            "default": "Azerty_1234"

          },
          {
            "name": "status",
            "description": "Le statut de l'utilisateur (actif ou inactif)",
            "in": "formData",
            "type": "boolean",
            "required": true,
            "schema": {
              "type": "boolean",
              "enum": [true, false]
            },
            "default": true

          },
          {
            "name": "role",
            "description": "Le role de l'utilisateur (user ou admin)",
            "in": "formData",
            "type": "string",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["user", "admin"]
            },
            "default": "user"

          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "tags": ["users"],
        "summary": "Modifier le compte d'un utilisateur",
        "description": "Permet de modifier les informations de l'utilisateur",
        "parameters": [
            {
              "name": "id",
              "in": "formData",
              "type": "integer",
              "required": true,
              "default": 1
            },
            {
              "name": "firstname",
              "in": "formData",
              "type": "string",
              "required": false
            },
            {
              "name": "lastname",
              "in": "formData",
              "type": "string",
              "required": false
            },
            {
              "name": "email",
              "in": "formData",
              "type": "string",
              "required": false
            },
            {
              "name": "description",
              "in": "formData",
              "type": "string",
              "required": false
            }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/routes": {
      "get": {
        "tags": ["routes"],
        "summary": "Obtenir tous les trajets",
        "description": "Permet d'obtenir tous les trajets existant",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/route": {
      "post": {
        "tags": ["routes"],
        "summary": "Obtenir un trajets via l'id",
        "description": "Permet de trouver un trajets via son id",
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/route/create": {
      "post": {
        "tags": ["routes"],
        "summary": "Créer un trajet",
        "description": "Permet de créer un trajet",
        "parameters": [
          {
            "name": "arrival_city_id",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "departure_city_id",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "departure_time",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2022-11-11 08:00:00"
          },
          {
            "name": "arrival_time",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2022-11-11 08:30:00"
          },
          {
            "name": "departure_date",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2022-11-11"
          },
          {
            "name": "vehicules_id",
            "in": "formData",
            "type": "string",
            "required": true,
            "default": "2"
          },
          {
            "name": "available_seats",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "remaining_seats",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/route/update": {
      "put": {
        "tags": ["routes"],
        "summary": "Obtenir les trajets via l'id de la ville de depart ou d'arrivée",
        "description": "Permet de trouver tout les trajets ayant comme ville d'arrivée ou de départ ou les deux la ville indiquée",
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "type": "integer",
            "required": true
          },
          {
            "name": "user_id",
            "in": "formData",
            "type": "integer",
            "required": false
          },
          {
            "name": "arrival_city_id",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "departure_city_id",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "departure_time",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "arrival_time",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "departure_date",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "available_seats",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "remaining_seats",
            "in": "formData",
            "type": "string",
            "required": false
          },
          {
            "name": "statuts",
            "in": "formData",
            "schema": {
              "type": "boolean",
              "enum": [true, false]
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/usersHasRoutes": {
      "get": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir tous les utilisateurs participant à un trajets",
        "description": "Permet d'obtenir tous les utilisateurs participant à un trajets",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/userHasRoute/route": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir les utilisateurs participant à un trajet",
        "description": "Permet de trouver les utilisateurs participant à un trajet",
        "parameters": [
          {
            "name": "route_id",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/userHasRoute/user": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir les trajets d'un utilisateur",
        "description": "Permet d'obtenir les trajets d'un utilisateur",
        "parameters": [
          {
            "name": "user_id",
            "in": "formData",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/userHasRoute/user/route": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Voir si l'utilisateur participe à un trajet",
        "description": "Permet de voir si l'utilisateur participe à un trajet",
        "parameters": [
          {
            "name": "route_id",
            "in": "formData",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/userHasRoute/create": {
      "post": {
        "tags": ["users_has_routes"],
        "summary": "Obtenir les trajets via l'id de la ville de depart ou d'arrivée",
        "description": "Permet de trouver tout les trajets ayant comme ville d'arrivée ou de départ ou les deux la ville indiquée",
        "parameters": [
          {
            "name": "user_id",
            "in": "formData",
            "type": "integer",
            "required": true
          },
          {
            "name": "route_id",
            "in": "formData",
            "type": "integer",
            "required": true
          },
          {
            "name": "status_notice",
            "in": "formData",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/userHasRoute/update": {
      "put": {
        "tags": ["users_has_routes"],
        "summary": "Modifier le participant ou le trajet",
        "description": "Permet de modifier le participant ou le trajet",
        "parameters": [
          {
            "name": "user_id",
            "in": "formData",
            "type": "integer",
            "required": false
          },
          {
            "name": "route_id",
            "in": "formData",
            "type": "integer",
            "required": false
          },
          {
            "name": "status_notice",
            "in": "formData",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}